+++ contracts/handlers/gelatov2/HGelatoV2LimitOrder.sol
@@ -9,21 +9,25 @@ contract HGelatoV2LimitOrder is HandlerBase {
     using SafeERC20 for IERC20;
     using SafeMath for uint256;
 
-    // prettier-ignore
-    address public constant GELATO_PINE = 0x36049D479A97CdE1fC6E2a5D2caE30B666Ebf92B;
-    // prettier-ignore
-    address public constant GELATO_LIMIT_ORDER_MODULE = 0x037fc8e71445910e1E0bBb2a0896d5e9A7485318;
+    address public immutable GELATO_PINE;
+    address public immutable GELATO_LIMIT_ORDER_MODULE;
+
+    constructor(address _gelatoPine, address _module) public {
+        GELATO_PINE = _gelatoPine;
+        GELATO_LIMIT_ORDER_MODULE = _module;
+    }
 
     function getContractName() public pure override returns (string memory) {
         return "HGelatoV2LimitOrder";
     }
 
     function placeLimitOrder(
-        address inToken,
-        address outToken,
         uint256 value,
-        uint256 minimumReturn,
+        address module, // unused, only for gelato graph bot
+        address inToken,
+        address payable owner, // unused, only for gelato graph bot
         address _witness,
+        bytes calldata limitOrderData,
         bytes32 _secret
     ) external payable {
         // Transfer funds inside Furu Proxy => Will be done by a separate Handler
@@ -41,7 +45,7 @@ contract HGelatoV2LimitOrder is HandlerBase {
                         inToken,
                         payable(_getSender()),
                         _witness,
-                        abi.encode(outToken, minimumReturn),
+                        limitOrderData,
                         _secret
                     )
                 )
@@ -51,25 +55,21 @@ contract HGelatoV2LimitOrder is HandlerBase {
                 _revertMsg("placeLimitOrder");
             }
         } else {
-            (bool success, bytes memory ret) =
-                inToken.call(
-                    IGelatoPineCore(GELATO_PINE).encodeTokenOrder(
+            try
+                IERC20(inToken).transfer(
+                    IGelatoPineCore(GELATO_PINE).vaultOfOrder(
                         GELATO_LIMIT_ORDER_MODULE,
-                        IERC20(inToken),
+                        inToken,
                         payable(_getSender()),
                         _witness,
-                        abi.encode(outToken, minimumReturn),
-                        _secret,
-                        value
-                    )
-                );
-
-            if (!success) {
-                if (ret.length < 68) _revertMsg("placeLimitOrder");
-                assembly {
-                    ret := add(ret, 0x04)
-                }
-                _revertMsg("placeLimitOrder", abi.decode(ret, (string)));
+                        limitOrderData
+                    ),
+                    value
+                )
+            {} catch Error(string memory reason) {
+                _revertMsg("placeLimitOrder", reason);
+            } catch {
+                _revertMsg("placeLimitOrder");
             }
         }
     }