+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [1.6.0] - 2021-06-15
+
 ### Added
 
 - Support Polygon token bridge.
@@ -13,6 +15,8 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Archive Uniswap and Mooniswap handlers.
+- Change post process handler checking flow.
+- Check proxy instead of sender for ban.
 
 ## [1.5.1] - 2021-05-05